Implemented powerups. Starting with a blast-shield.
Implemented a basic powerup framework, the client registers/sets the powerup type for the current player which in turn sets the powerup for that player on the server (shield_powerup_model) with everyone.now callbacks for activate and deactivate.

When the client activates the powerup, the serverside powerup runs it's activate callback and activates that powerup on all client browsers.

This framework should make adding new powerups trivial. Also, since there are nice hooks, implementing it on special-ball-pull should also be easy.

We should be able to use the servside code to handle the actual effects of the powerups.